Split package checksum helpers (#270)
